seperate retire function
